requirements.txt
adrf==0.1.4
bayoo-docx==0.2.20
camelot-py==0.11.0
datasets==2.4.0
django-filter==23.2
django-rest-swagger==2.2.0
django==4.2.1
djangorestframework==3.14.0
drf_spectacular==0.26.4
drf_yasg==1.21.7
httpx==0.27.0
imutils==0.5.4
inflect==7.0.0
lxml==4.9.2
ndjson==0.3.1
nltk==3.8
num2words==0.5.13
opencv-python-headless==4.8.1.78
pandas==2.2.0
pdf2image==1.16.3
Pillow==9.5.0
pymupdf==1.22.3
pytesseract==0.3.10
python-docx==0.8.11
roman_arabic_numerals==0.8.0
sentence_transformers==2.2.2
spacy==3.5.3
torch==2.0.1
transformers==4.35.2
xlsxwriter==3.1.9